+++ packages/legend-extension-dsl-data-space/src/__lib__/query/DSL_DataSpace_LegendQueryNavigation.ts
@@ -101,7 +101,7 @@ export const generateDataSpaceQueryCreatorRoute = (
     }),
   );
 
-export const generateDataSpaceTemplateQueryViewerRoute = (
+export const generateDataSpaceTemplateQueryCreatorRoute = (
   groupId: string,
   artifactId: string,
   versionId: string,
