+++ packages/legend-extension-dsl-data-space/src/components/query/DataSpaceQueryBuilder.tsx
@@ -27,12 +27,6 @@ import {
   CheckIcon,
   MenuContentItemLabel,
   SearchIcon,
-  FilterIcon,
-  PanelHeader,
-  BasePopover,
-  TagIcon,
-  ClickAwayListener,
-  ShareIcon,
 } from '@finos/legend-art';
 import { observer } from 'mobx-react-lite';
 import { useApplicationStore } from '@finos/legend-application';
@@ -53,18 +47,16 @@ import {
 } from '@finos/legend-graph';
 import type { DataSpaceInfo } from '../../stores/query/DataSpaceInfo.js';
 import { generateGAVCoordinates } from '@finos/legend-storage';
-import { useEffect, useMemo, useRef, useState } from 'react';
+import { useEffect, useMemo, useState } from 'react';
 import { debounce, guaranteeType } from '@finos/legend-shared';
 import { flowResult } from 'mobx';
 import {
   DataSpace,
-  DataSpaceExecutableTemplate,
   type DataSpaceExecutionContext,
 } from '../../graph/metamodel/pure/model/packageableElements/dataSpace/DSL_DataSpace_DataSpace.js';
 import { DataSpaceIcon } from '../DSL_DataSpace_Icon.js';
 import { DataSpaceAdvancedSearchModal } from './DataSpaceAdvancedSearchModal.js';
 import type { EditorStore } from '@finos/legend-application-studio';
-import { generateDataSpaceTemplateQueryViewerRoute } from '../../__lib__/query/DSL_DataSpace_LegendQueryNavigation.js';
 
 export type DataSpaceOption = {
   label: string;
@@ -397,168 +389,6 @@ export const renderDataSpaceQueryBuilderSetupPanelContent = (
   queryBuilderState: DataSpaceQueryBuilderState,
 ): React.ReactNode => (
   <DataSpaceQueryBuilderSetupPanelContent
-    queryBuilderState={queryBuilderState}
-  />
-);
-
-const DataSpaceTemplateQueryDialog = observer(
-  (props: {
-    triggerElement: HTMLElement | null;
-    queryBuilderState: DataSpaceQueryBuilderState;
-    templateQueries: DataSpaceExecutableTemplate[];
-  }) => {
-    const { triggerElement, queryBuilderState, templateQueries } = props;
-    const applicationStore = useApplicationStore();
-    const handleClose = (): void => {
-      queryBuilderState.setTemplateQueryDialogOpen(false);
-    };
-    const loadQuery = (template: DataSpaceExecutableTemplate): void => {
-      const executionContext =
-        queryBuilderState.dataSpace.executionContexts.find(
-          (c) => c.name === template.executionContextKey,
-        );
-      if (
-        executionContext &&
-        executionContext.hashCode !==
-          queryBuilderState.executionContext.hashCode
-      ) {
-        queryBuilderState.setExecutionContext(executionContext);
-        queryBuilderState.propagateExecutionContextChange(executionContext);
-        queryBuilderState.initializeWithQuery(template.query);
-        queryBuilderState.onExecutionContextChange?.(executionContext);
-      } else {
-        queryBuilderState.initializeWithQuery(template.query);
-      }
-      handleClose();
-    };
-
-    const shareTemplateQuery = (
-      template: DataSpaceExecutableTemplate,
-    ): void => {
-      if (queryBuilderState.projectInfo?.groupId) {
-        applicationStore.navigationService.navigator.visitAddress(
-          applicationStore.navigationService.navigator.generateAddress(
-            generateDataSpaceTemplateQueryViewerRoute(
-              queryBuilderState.projectInfo.groupId,
-              queryBuilderState.projectInfo.artifactId,
-              queryBuilderState.projectInfo.versionId,
-              queryBuilderState.dataSpace.path,
-              template.title,
-            ),
-          ),
-        );
-      }
-    };
-
-    return (
-      <ClickAwayListener onClickAway={handleClose}>
-        <div>
-          <BasePopover
-            open={queryBuilderState.isTemplateQueryDialogOpen}
-            PaperProps={{
-              classes: {
-                root: '"query-builder__data-space__template-query-panel__container__root',
-              },
-            }}
-            className="query-builder__data-space__template-query-panel__container"
-            onClose={handleClose}
-            anchorEl={triggerElement}
-            anchorOrigin={{
-              vertical: 'bottom',
-              horizontal: 'left',
-            }}
-            transformOrigin={{
-              vertical: 'top',
-              horizontal: 'center',
-            }}
-          >
-            <div className="query-builder__data-space__template-query-panel">
-              <div className="query-builder__data-space__template-query-panel__header">
-                Curated Template Queries
-              </div>
-              {templateQueries.map((query) => (
-                <div
-                  key={query.title}
-                  className="query-builder__data-space__template-query-panel__query"
-                >
-                  <TagIcon className="query-builder__data-space__template-query-panel__query__icon" />
-                  <button
-                    className="query-builder__data-space__template-query-panel__query__entry"
-                    title="click to load template query"
-                    onClick={() => loadQuery(query)}
-                  >
-                    <div className="query-builder__data-space__template-query-panel__query__entry__content">
-                      <div className="query-builder__data-space__template-query-panel__query__entry__content__title">
-                        {query.title}
-                      </div>
-                      {query.description && (
-                        <div className="query-builder__data-space__template-query-panel__query__entry__content__description">
-                          {query.description}
-                        </div>
-                      )}
-                    </div>
-                  </button>
-                  <button
-                    className="query-builder__data-space__template-query-panel__query__share"
-                    title="Share..."
-                    onClick={() => shareTemplateQuery(query)}
-                  >
-                    <ShareIcon />
-                    <div className="query-builder__data-space__template-query-panel__query__share__label">
-                      Share
-                    </div>
-                  </button>
-                </div>
-              ))}
-            </div>
-          </BasePopover>
-        </div>
-      </ClickAwayListener>
-    );
-  },
-);
-
-const DataSpaceQueryBuilderTemplateQueryPanel = observer(
-  (props: { queryBuilderState: DataSpaceQueryBuilderState }) => {
-    const { queryBuilderState } = props;
-    const templateQueryButtonRef = useRef<HTMLButtonElement>(null);
-    const templateQueries = queryBuilderState.dataSpace.executables?.filter(
-      (e) => e instanceof DataSpaceExecutableTemplate,
-    ) as DataSpaceExecutableTemplate[];
-
-    const showTemplateQueries = (): void => {
-      queryBuilderState.setTemplateQueryDialogOpen(true);
-    };
-
-    return (
-      <PanelHeader className="query-builder__data-space__template-query">
-        <div className="query-builder__data-space__template-query__title">
-          <FilterIcon />
-        </div>
-        <button
-          className="query-builder__data-space__template-query__btn"
-          ref={templateQueryButtonRef}
-          disabled={templateQueries.length <= 0}
-          onClick={showTemplateQueries}
-        >
-          Template ( {templateQueries.length} )
-        </button>
-        {queryBuilderState.isTemplateQueryDialogOpen && (
-          <DataSpaceTemplateQueryDialog
-            triggerElement={templateQueryButtonRef.current}
-            queryBuilderState={queryBuilderState}
-            templateQueries={templateQueries}
-          />
-        )}
-      </PanelHeader>
-    );
-  },
-);
-
-export const renderDataSpaceQueryBuilderTemplateQueryPanelContent = (
-  queryBuilderState: DataSpaceQueryBuilderState,
-): React.ReactNode => (
-  <DataSpaceQueryBuilderTemplateQueryPanel
     queryBuilderState={queryBuilderState}
   />
 );
